Move original source into gem lib
